refactor: Chat CoreData, Socket Repository 리팩토링 [#30]
