+++ SweetHome/Data/Repositories/ChatCoreDataRepositoryImpl.swift
@@ -1,5 +1,5 @@
 //
-//  ChatCoreDataRepository.swift
+//  ChatCoreDataRepositoryImpl.swift
 //  SweetHome
 //
 //  Created by 김민호 on 8/26/25.
@@ -9,37 +9,13 @@ import CoreData
 import Foundation
 import RxSwift
 
-protocol ChatLocalRepository {
-    // MARK: - Chat Rooms
-    func saveChatRoom(_ chatRoom: ChatRoom) -> Observable<Void>
-    func fetchChatRooms() -> Observable<[ChatRoom]>
-    func updateChatRoomUnreadCount(roomId: String, count: Int) -> Observable<Void>
-    func deleteChatRoom(roomId: String) -> Observable<Void>
-    
-    // MARK: - Chat Messages
-    func saveChatMessage(_ message: LastChat) -> Observable<Void>
-    func saveChatMessages(_ messages: [LastChat]) -> Observable<Void>
-    func fetchChatMessages(for roomId: String) -> Observable<[LastChat]>
-    func fetchChatMessages(for roomId: String, limit: Int) -> Observable<[LastChat]>
-    func updateMessageReadStatus(chatId: String, isRead: Bool) -> Observable<Void>
-    func deleteChatMessages(for roomId: String) -> Observable<Void>
-    
-    // MARK: - Unread Count Management
-    func incrementUnreadCount(for roomId: String) -> Observable<Void>
-    func resetUnreadCount(for roomId: String) -> Observable<Void>
-    
-    // MARK: - Sync Management
-    func getLastMessageDate(for roomId: String) -> Observable<Date?>
-    func markMessagesAsRead(for roomId: String, upTo lastReadChatId: String) -> Observable<Void>
-    func getLatestMessageForRoom(roomId: String) -> Observable<LastChat?>
-}
-
-class ChatCoreDataRepository: ChatLocalRepository {
+/// - CoreData를 사용하는 ChatLocalRepository 구현체
+final class ChatCoreDataRepositoryImpl: ChatLocalRepository {
     private let coreDataStack = CoreDataStack.shared
 }
 
 // MARK: - 채팅방 관련 메서드
-extension ChatCoreDataRepository {
+extension ChatCoreDataRepositoryImpl {
     func saveChatRoom(_ chatRoom: ChatRoom) -> Observable<Void> {
         return Observable.create { [weak self] observer in
             guard let self = self else {
@@ -187,7 +163,7 @@ extension ChatCoreDataRepository {
     }
 }
 // MARK: - 채팅 메세지 관련 메서드
-extension ChatCoreDataRepository {
+extension ChatCoreDataRepositoryImpl {
     func saveChatMessage(_ message: LastChat) -> Observable<Void> {
         return Observable.create { [weak self] observer in
             guard let self else {
@@ -376,7 +352,7 @@ extension ChatCoreDataRepository {
     }
 }
 // MARK: - 안읽은 메세지 관련 메서드
-extension ChatCoreDataRepository {
+extension ChatCoreDataRepositoryImpl {
     func incrementUnreadCount(for roomId: String) -> Observable<Void> {
         return Observable.create { [weak self] observer in
             guard let self = self else {
@@ -418,7 +394,7 @@ extension ChatCoreDataRepository {
     }
 }
 // MARK: - 동기화 및 읽음처리 관련 메서드
-extension ChatCoreDataRepository {
+extension ChatCoreDataRepositoryImpl {
     func getLastMessageDate(for roomId: String) -> Observable<Date?> {
         return Observable.create { [weak self] observer in
             guard let self = self else {
@@ -492,7 +468,7 @@ extension ChatCoreDataRepository {
     }
 }
 // MARK: - Private Helpers
-extension ChatCoreDataRepository {
+extension ChatCoreDataRepositoryImpl {
     private func fetchLastMessage(for roomId: String) -> LastChat? {
         let context = coreDataStack.context
         let fetchRequest: NSFetchRequest<SweetHome.CDChatMessage> = SweetHome.CDChatMessage.fetchRequest()
